MAIN/requirements.txt
SQLAlchemy == 1.4
uvicorn==0.29.0
fastapi == 0.103.2
psycopg2-binary == 2.9.9
python-dotenv == 1.0.1

